Add files via upload
🌟 Netflix Clone Homepage 🎬

Welcome to my GitHub repository featuring a stunning Netflix Clone Homepage! 🚀

Using HTML and CSS, I've meticulously crafted a pixel-perfect replica of the popular streaming platform's homepage. 🎉 The sleek design, dynamic layout, and attention to detail make this project a standout showcase of my web development skills. 💻

Feel free to explore the code, clone the repository, and even contribute if you'd like! 🤝 Let's collaborate and take this project to new heights together. 💡

Check out the live demo here and immerse yourself in the world of seamless user experience. 📺 Don't forget to give it a ⭐️ if you find it interesting!

#HTML #CSS #WebDevelopment #NetflixClone #FrontendDevelopment